src/opdracht1/GearboxTest.java
package opdracht1;

import static org.junit.Assert.assertEquals;

import org.junit.Test;

public class GearboxTest{
	
	@Test
	public void testGearboxConstructorAndWheelSpeed() {
		Gearbox mcLaren = new Gearbox(3);
		assertEquals(0.0, mcLaren.wheelSpeed(1000), 0.01);
		mcLaren.operateClutch(true);
		mcLaren.changeGear(1);
		mcLaren.operateClutch(false);
		assertEquals(5300.0, mcLaren.wheelSpeed(1000), 0.01);
		mcLaren.operateClutch(true);
		mcLaren.changeGear(2);
		mcLaren.operateClutch(false);
		assertEquals(10600.0, mcLaren.wheelSpeed(1000), 0.01);
		mcLaren.operateClutch(true);
		mcLaren.changeGear(3);
		mcLaren.operateClutch(false);
		assertEquals(15900.0, mcLaren.wheelSpeed(1000), 0.01);
	}
	
	@Test
	public void changeGearNotAllowedWhenClutchIsInFalse() {
		Gearbox mcLaren = new Gearbox(3);
		assertEquals(0.0, mcLaren.wheelSpeed(1000), 0.01);
		mcLaren.operateClutch(true);
		mcLaren.changeGear(1);
		mcLaren.operateClutch(false);
		
		mcLaren.changeGear(2);
		assertEquals(0.0, mcLaren.wheelSpeed(1000), 0.01);
	}
	
	@Test
	public void changeGearNotAllowedWhenNewGearNegative() {
		Gearbox mcLaren = new Gearbox(3);
		assertEquals(0.0, mcLaren.wheelSpeed(1000), 0.01);
		mcLaren.operateClutch(true);
		mcLaren.changeGear(1);
		mcLaren.operateClutch(false);
		
		mcLaren.changeGear(-1);
		assertEquals(0.0, mcLaren.wheelSpeed(1000), 0.01);
	}
	
	@Test
	public void changeGearNotAllowedWhenNewGearGreaterThanMaxGears() {
		Gearbox mcLaren = new Gearbox(3);
		assertEquals(0.0, mcLaren.wheelSpeed(1000), 0.01);
		mcLaren.operateClutch(true);
		mcLaren.changeGear(1);

		
		mcLaren.changeGear(4);
		assertEquals(0.0, mcLaren.wheelSpeed(1000), 0.01);
	}
	
	
}
